requirements.txt
#存放库
#存放后要激活
colorama
requests
python-dotenv
openai